services/msautogen/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
arxiv==2.1.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker==7.0.0
feedparser==6.0.10
FLAML==2.1.2
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
numpy==1.26.4
openai==1.14.2
packaging==24.0
pandas==2.2.1
pika==1.3.2
pyautogen==0.2.20
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
termcolor==2.4.0
tiktoken==0.6.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
